pyproject.toml
[project]
name = "agentfromscratch"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13.5"
dependencies = [
    "google-genai>=1.59.0",
    "openai>=2.15.0",
    "python-dotenv>=1.2.1",
    "strands-agents>=1.25.0",
    "strands-agents-tools>=0.2.20",
]
